Refactor: move LogId to sub folder log_id/
